2019/04/07 - fix data types - most of the doubles should be decimal for dollar amount properties

+++ CoderGirl-SalesList/ISalesRecordAnalyzer.cs
@@ -25,14 +25,14 @@ namespace CoderGirl_SalesList
         /// </summary>
         /// <param name="salesRecords"></param>
         /// <returns></returns>
-        double GetMaxProfit(List<SalesRecord> salesRecords);
+        decimal GetMaxProfit(List<SalesRecord> salesRecords);
 
         /// <summary>
         /// Returns the Sum of the TotalRevenue of all Sales Records
         /// </summary>
         /// <param name="salesRecords"></param>
         /// <returns></returns>
-        double GetTotalRevenue(List<SalesRecord> salesRecords);
+        decimal GetTotalRevenue(List<SalesRecord> salesRecords);
 
         /// <summary>
         /// Returns list of SalesRecords ordered by ShipDate in ascending order
